# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/finetune": {
            "get": {
                "description": "list finetunes",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Finetune"
                ],
                "summary": "List",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.UserFinetunesDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            },
            "post": {
                "description": "create finetune",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Finetune"
                ],
                "summary": "Create",
                "parameters": [
                    {
                        "description": "body of creating finetune",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.FinetuneCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/controller.finetuneCreateResp"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/finetune/ws": {
            "get": {
                "description": "watch finetunes",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Finetune"
                ],
                "summary": "WatchFinetunes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.FinetuneSummaryDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/finetune/{id}": {
            "put": {
                "description": "terminate finetune",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Finetune"
                ],
                "summary": "Terminate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "finetune id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            },
            "delete": {
                "description": "delete finetune",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Finetune"
                ],
                "summary": "Delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "finetune id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/finetune/{id}/log": {
            "get": {
                "description": "download finetune log",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Finetune"
                ],
                "summary": "Log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "finetune id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.finetuneLog"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
        "/v1/finetune/{id}/log/ws": {
            "get": {
                "description": "watch single finetune",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Finetune"
                ],
                "summary": "WatchSingle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "finetune id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controller.finetuneLog"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
        "/v1/user/follower/{account}": {
            "get": {
                "description": "list followers",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Follower"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "the account the followers belong to",
                        "name": "account",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.FollowsDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "system_error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.FinetuneSummaryDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_done": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "app.FollowDTO": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "avatar_id": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "is_follower": {
                    "type": "boolean"
                }
            }
        },
        "app.FollowsDTO": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.FollowDTO"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "app.UserFinetunesDTO": {
            "type": "object",
            "properties": {
                "datas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/app.FinetuneSummaryDTO"
                    }
                },
                "expiry": {
                    "type": "integer"
                }
            }
        },
...
        "controller.FinetuneCreateRequest": {
            "type": "object",
            "properties": {
                "hyperparameter": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controller.KeyValue"
                    }
                },
                "model": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "task": {
                    "type": "string"
                }
            }
        },
...
        "controller.KeyValue": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "controller.finetuneCreateResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "controller.finetuneLog": {
            "type": "object",
            "properties": {
                "log": {
                    "type": "string"
                }
            }
        },
...
    }
}